Extra 10% gain in performance due to inline of nxtpcb() and nxtpcw(), removing unnecessary function calls in the execution of opcodes. now, peekb/w() is directly called by the opcodes.
A slight simplification had to be made for the inline be possible. This simplification does not correct a possible overflow of the _PC value after it is incremented. This overflow could occur in case _PC=0xffff and one of the above functions is called, but (1) this situation never arised in the tests and (2) it *seems* the opcodes already take extra care to correct any overflown _PC that might have been received (they all do an AND 0xffff whenever _PC is used). 

Therefore, it seems that this simplification is worthwhile the performance gain obtained.


git-svn-id: http://svn.code.sf.net/p/jsmsx/code@22 f0662447-c21b-0410-9b8f-8ded6cf39f67